+++ HVLib/Types/Base/HVMinute.m
@@ -20,11 +20,11 @@
 
 @implementation HVMinute
 
--(int) min
+-(NSInteger) min
 {
     return 0;
 }
--(int) max
+-(NSInteger) max
 {
     return 59;
 }
